vps/summi_worker/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4

